+++ src/main.js
@@ -6,8 +6,7 @@ class AcodeAceLinters {
 		this.baseUrl = this.baseUrl.endsWith("/")
 			? this.baseUrl
 			: this.baseUrl + "/";
-		const workerBlob = this.createWorkerBlob();
-		const workerUrl = URL.createObjectURL(workerBlob);
+		const workerUrl = this.baseUrl + "worker.js";
 
 		this.worker = new Worker(workerUrl);
 		this.provider = LanguageProvider.create(this.worker);
@@ -34,100 +33,6 @@ class AcodeAceLinters {
 		});
 
 		window.toast("Ace Linters: Loaded successfully");
-	}
-
-	createWorkerBlob() {
-		// This list should match the files we copied to 'dist' in esbuild config
-		const services = [
-			{
-				name: "json",
-				script: "json-service.js",
-				className: "JsonService",
-				modes: "json|json5",
-			},
-			{
-				name: "html",
-				script: "html-service.js",
-				className: "HtmlService",
-				modes: "html",
-			},
-			{
-				name: "css",
-				script: "css-service.js",
-				className: "CssService",
-				modes: "css",
-			},
-			{
-				name: "less",
-				script: "css-service.js",
-				className: "CssService",
-				modes: "less",
-			},
-			{
-				name: "scss",
-				script: "css-service.js",
-				className: "CssService",
-				modes: "scss",
-			},
-			{
-				name: "typescript",
-				script: "typescript-service.js",
-				className: "TypescriptService",
-				modes: "typescript|tsx|javascript|jsx",
-			},
-			{
-				name: "lua",
-				script: "ace-lua-linter.js",
-				className: "AceLuaLinter",
-				modes: "lua",
-			},
-			{
-				name: "yaml",
-				script: "yaml-service.js",
-				className: "YamlService",
-				modes: "yaml",
-			},
-			{
-				name: "xml",
-				script: "xml-service.js",
-				className: "XmlService",
-				modes: "xml",
-			},
-			{
-				name: "php",
-				script: "php-service.js",
-				className: "PhpService",
-				modes: "php",
-			},
-		];
-
-		// We construct the worker code as a string.
-		// Note: We use `importScripts` which works inside the Blob worker because we pass the absolute `baseUrl` to locate the files.
-		const workerCode = `
-            try {
-                importScripts("${this.baseUrl}service-manager.js");
-                const manager = new ServiceManager(self);
-                
-                ${services
-									.map(
-										(s) => `
-                    manager.registerService("${s.name}", {
-                        module: () => {
-                            importScripts("${this.baseUrl}${s.script}");
-                            return { ${s.className} };
-                        },
-                        className: "${s.className}",
-                        modes: "${s.modes}"
-                    });
-                `,
-									)
-									.join("\n")}
-            } catch (e) {
-                console.error("Worker Error:", e);
-            }
-        `;
-
-		return new Blob([workerCode], { type: "application/javascript" });
 	}
 
 	async destroy() {